requirements.txt
# requirements.txt
streamlit>=1.28.0
requests>=2.31.0
openai>=1.3.0
psycopg2-binary>=2.9.0